requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
numpy==1.26.0
matplotlib==3.8.0
httpx==0.26.0
python-multipart==0.0.9
pandas==2.2.0
seaborn==0.13.0